+++ src/ta_func/ta_APO.c
@@ -74,7 +74,7 @@ TA_LIB_API int TA_APO_Lookback( int optInFastPeriod, int optInSlowPeriod, TA_MAT
    else if( (int)optInSlowPeriod < 2 || (int)optInSlowPeriod > 100000 )
       return -1;
    if( (int)optInMAType == (int)0x80000000 )
-      optInMAType = 0;
+      optInMAType = 1;
    /* The slow MA is the key factor determining the lookback period. */
    return TA_MA_Lookback(max(optInSlowPeriod,optInFastPeriod),optInMAType);
 }
@@ -113,7 +113,7 @@ TA_LIB_API TA_RetCode TA_APO( int    startIdx,
    else if( (int)optInSlowPeriod < 2 || (int)optInSlowPeriod > 100000 )
       return TA_BAD_PARAM;
    if( (int)optInMAType == (int)0x80000000 )
-      optInMAType = 0;
+      optInMAType = 1;
    if( !outReal )
       return TA_BAD_PARAM;
 
@@ -245,7 +245,7 @@ TA_RetCode TA_S_APO( int    startIdx,
    else if( (int)optInSlowPeriod < 2 || (int)optInSlowPeriod > 100000 )
       return TA_BAD_PARAM;
    if( (int)optInMAType == (int)0x80000000 )
-      optInMAType = 0;
+      optInMAType = 1;
    if( !outReal )
       return TA_BAD_PARAM;
 
@@ -392,7 +392,7 @@ TA_RetCode TA_APO_OpenInternal( struct TA_APO_Stream **stream, const double inRe
    else if( (int)optInSlowPeriod < 2 || (int)optInSlowPeriod > 100000 )
       return TA_BAD_PARAM;
    if( (int)optInMAType == (int)0x80000000 )
-      optInMAType = 0;
+      optInMAType = 1;
 
    endIdx = historyLen - 1;
    dummyBegIdx = 0;