+++ src/ta_func/ta_PPO.c
@@ -73,7 +73,7 @@ TA_LIB_API int TA_PPO_Lookback( int optInFastPeriod, int optInSlowPeriod, TA_MAT
    else if( (int)optInSlowPeriod < 2 || (int)optInSlowPeriod > 100000 )
       return -1;
    if( (int)optInMAType == (int)0x80000000 )
-      optInMAType = 0;
+      optInMAType = 1;
    /* Lookback is driven by the slowest MA. */
    return TA_MA_Lookback(max(optInSlowPeriod,optInFastPeriod),optInMAType);
 }
@@ -113,7 +113,7 @@ TA_LIB_API TA_RetCode TA_PPO( int    startIdx,
    else if( (int)optInSlowPeriod < 2 || (int)optInSlowPeriod > 100000 )
       return TA_BAD_PARAM;
    if( (int)optInMAType == (int)0x80000000 )
-      optInMAType = 0;
+      optInMAType = 1;
    if( !outReal )
       return TA_BAD_PARAM;
 
@@ -261,7 +261,7 @@ TA_RetCode TA_S_PPO( int    startIdx,
    else if( (int)optInSlowPeriod < 2 || (int)optInSlowPeriod > 100000 )
       return TA_BAD_PARAM;
    if( (int)optInMAType == (int)0x80000000 )
-      optInMAType = 0;
+      optInMAType = 1;
    if( !outReal )
       return TA_BAD_PARAM;
 
@@ -431,7 +431,7 @@ TA_RetCode TA_PPO_OpenInternal( struct TA_PPO_Stream **stream, const double inRe
    else if( (int)optInSlowPeriod < 2 || (int)optInSlowPeriod > 100000 )
       return TA_BAD_PARAM;
    if( (int)optInMAType == (int)0x80000000 )
-      optInMAType = 0;
+      optInMAType = 1;
 
    endIdx = historyLen - 1;
    dummyBegIdx = 0;